Functions
The way functions are implemented in the tutorial is pretty messy.
The syntax is a mischmasch of expressions and values. Functions
are represented partly by their semantics so the syntax depends on
the semantics. This means that I could no longer have syntax and
semantics in separate modules. So I caved and inlined Expr.hs into
Eval.hs. Sad.